Valore com virugla

+++ Models/ComprasViewModel.cs
@@ -13,8 +13,8 @@ namespace TesteUGBMVC.Models
         public DateTime DataSolicitada { get; set; }
         public DateTime DataPrevisaoEntregaProduto { get; set; }
         public TipoDoProdutoEnum TipoDoProduto { get; set; }
-        public int ValorUnitarioDoProduto { get; set; }
-        public int ValorTotal { get; set; } // Soma do valor unitario + quantidade
+        public double ValorUnitarioDoProduto { get; set; }
+        public double ValorTotal { get; set; } // Soma do valor unitario + quantidade
         public long NumeroNotaFiscalProduto { get; set; }
         public long CodigoEAN { get; set; } // Código de Barras do Produto
     }
